Refactor dynamic_hmc out of hmc.py
