+++ blackjax/mcmc/hmc.py
@@ -26,14 +26,10 @@ from blackjax.types import Array, ArrayLikeTree, ArrayTree, PRNGKey
 
 __all__ = [
     "HMCState",
-    "DynamicHMCState",
     "HMCInfo",
     "init",
-    "init_dynamic",
     "build_kernel",
-    "build_dynamic_kernel",
     "hmc",
-    "dynamic_hmc",
 ]
 
 
@@ -49,20 +45,6 @@ class HMCState(NamedTuple):
     position: ArrayTree
     logdensity: float
     logdensity_grad: ArrayTree
-
-
-class DynamicHMCState(NamedTuple):
-    """State of the dynamic HMC algorithm.
-
-    Adds a utility array for generating a pseudo or quasi-random sequence of
-    number of integration steps.
-
-    """
-
-    position: ArrayTree
-    logdensity: float
-    logdensity_grad: ArrayTree
-    random_generator_arg: Array
 
 
 class HMCInfo(NamedTuple):
@@ -106,13 +88,6 @@ class HMCInfo(NamedTuple):
 def init(position: ArrayLikeTree, logdensity_fn: Callable):
     logdensity, logdensity_grad = jax.value_and_grad(logdensity_fn)(position)
     return HMCState(position, logdensity, logdensity_grad)
-
-
-def init_dynamic(
-    position: ArrayLikeTree, logdensity_fn: Callable, random_generator_arg: Array
-):
-    logdensity, logdensity_grad = jax.value_and_grad(logdensity_fn)(position)
-    return DynamicHMCState(position, logdensity, logdensity_grad, random_generator_arg)
 
 
 def build_kernel(
@@ -172,70 +147,6 @@ def build_kernel(
         )
 
         return proposal, info
-
-    return kernel
-
-
-def build_dynamic_kernel(
-    integrator: Callable = integrators.velocity_verlet,
-    divergence_threshold: float = 1000,
-    next_random_arg_fn: Callable = lambda key: jax.random.split(key)[1],
-    integration_steps_fn: Callable = lambda key: jax.random.randint(key, (), 1, 10),
-):
-    """Build a Dynamic HMC kernel where the number of integration steps is chosen randomly.
-
-    Parameters
-    ----------
-    integrator
-        The symplectic integrator to use to integrate the Hamiltonian dynamics.
-    divergence_threshold
-        Value of the difference in energy above which we consider that the transition is divergent.
-    next_random_arg_fn
-        Function that generates the next `random_generator_arg` from its previous value.
-    integration_steps_fn
-        Function that generates the next pseudo or quasi-random number of integration steps in the
-        sequence, given the current `random_generator_arg`. Needs to return an `int`.
-
-    Returns
-    -------
-    A kernel that takes a rng_key and a Pytree that contains the current state
-    of the chain and that returns a new state of the chain along with
-    information about the transition.
-
-    """
-    hmc_base = build_kernel(integrator, divergence_threshold)
-
-    def kernel(
-        rng_key: PRNGKey,
-        state: DynamicHMCState,
-        logdensity_fn: Callable,
-        step_size: float,
-        inverse_mass_matrix: Array,
-        **integration_steps_kwargs,
-    ) -> tuple[DynamicHMCState, HMCInfo]:
-        """Generate a new sample with the HMC kernel."""
-        num_integration_steps = integration_steps_fn(
-            state.random_generator_arg, **integration_steps_kwargs
-        )
-        hmc_state = HMCState(state.position, state.logdensity, state.logdensity_grad)
-        hmc_proposal, info = hmc_base(
-            rng_key,
-            hmc_state,
-            logdensity_fn,
-            step_size,
-            inverse_mass_matrix,
-            num_integration_steps,
-        )
-        next_random_arg = next_random_arg_fn(state.random_generator_arg)
-        return (
-            DynamicHMCState(
-                hmc_proposal.position,
-                hmc_proposal.logdensity,
-                hmc_proposal.logdensity_grad,
-                next_random_arg,
-            ),
-            info,
-        )
 
     return kernel
 
@@ -332,69 +243,6 @@ class hmc:
             )
 
         return SamplingAlgorithm(init_fn, step_fn)
-
-
-class dynamic_hmc:
-    """Implements the (basic) user interface for the dynamic HMC kernel.
-
-    Parameters
-    ----------
-    logdensity_fn
-        The log-density function we wish to draw samples from.
-    step_size
-        The value to use for the step size in the symplectic integrator.
-    inverse_mass_matrix
-        The value to use for the inverse mass matrix when drawing a value for
-        the momentum and computing the kinetic energy.
-    divergence_threshold
-        The absolute value of the difference in energy between two states above
-        which we say that the transition is divergent. The default value is
-        commonly found in other libraries, and yet is arbitrary.
-    integrator
-        (algorithm parameter) The symplectic integrator to use to integrate the trajectory.
-    next_random_arg_fn
-        Function that generates the next `random_generator_arg` from its previous value.
-    integration_steps_fn
-        Function that generates the next pseudo or quasi-random number of integration steps in the
-        sequence, given the current `random_generator_arg`.
-
-
-    Returns
-    -------
-    A ``SamplingAlgorithm``.
-    """
-
-    init = staticmethod(init_dynamic)
-    build_kernel = staticmethod(build_dynamic_kernel)
-
-    def __new__(  # type: ignore[misc]
-        cls,
-        logdensity_fn: Callable,
-        step_size: float,
-        inverse_mass_matrix: Array,
-        *,
-        divergence_threshold: int = 1000,
-        integrator: Callable = integrators.velocity_verlet,
-        next_random_arg_fn: Callable = lambda key: jax.random.split(key)[1],
-        integration_steps_fn: Callable = lambda key: jax.random.randint(key, (), 1, 10),
-    ) -> SamplingAlgorithm:
-        kernel = cls.build_kernel(
-            integrator, divergence_threshold, next_random_arg_fn, integration_steps_fn
-        )
-
-        def init_fn(position: ArrayLikeTree, random_generator_arg: Array):
-            return cls.init(position, logdensity_fn, random_generator_arg)
-
-        def step_fn(rng_key: PRNGKey, state):
-            return kernel(
-                rng_key,
-                state,
-                logdensity_fn,
-                step_size,
-                inverse_mass_matrix,
-            )
-
-        return SamplingAlgorithm(init_fn, step_fn)  # type: ignore[arg-type]
 
 
 def hmc_proposal(
